+++ metatensor-core/CHANGELOG.md
@@ -26,6 +26,12 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
   called
 - Pin getrandom to make sure the code compiles with rustc 1.74
 
+#### Removed
+
+- `LabelsView` has been removed, and with it the following functions:
+  `Labels.is_view()`, `Labels.to_owned()`, `Labels.view()`, and
+  `Labels.__getitem__(list[str])`. We recomend using `Labels.column()` instead to access the values of individual dimensions of Labels.
+
 ## [Version 0.1.19](https://github.com/metatensor/metatensor/releases/tag/metatensor-core-v0.1.19) - 2025-12-11
 
 ### Fixed